+++ QuickFIXn/Store/FileStore.cs
@@ -269,10 +269,10 @@ public class FileStore : IMessageStore
     private void SetSeqNum()
     {
         _seqNumsFile.Seek(0, System.IO.SeekOrigin.Begin);
-        System.IO.StreamWriter writer = new System.IO.StreamWriter(_seqNumsFile);
-
-        writer.Write(NextSenderMsgSeqNum.ToString("D20") + " : " + NextTargetMsgSeqNum.ToString("D20") + "  ");
-        writer.Flush();
+        using (System.IO.StreamWriter writer = new System.IO.StreamWriter(_seqNumsFile, leaveOpen: true))
+        {
+            writer.Write(NextSenderMsgSeqNum.ToString("D20") + " : " + NextTargetMsgSeqNum.ToString("D20") + "  ");
+        }
     }
 
     public DateTime? CreationTime => _cache.CreationTime;